Bump version, integrate symphonia to README
