+++ examples/playlist.rs
@@ -1,5 +1,8 @@
+//! Analyze a folder recursively, and make a playlist out of the file
+//! provided by the user.
+//! How to use: ./playlist [-o file.m3u] [-a analysis.json] <folder> <file to start the playlist from>
 use anyhow::Result;
-use bliss_audio::decoder::ffmpeg::FFmpeg as Decoder;
+use bliss_audio::decoder::ffmpeg::FFmpegDecoder as Decoder;
 use bliss_audio::decoder::Decoder as DecoderTrait;
 use bliss_audio::playlist::{closest_to_songs, dedup_playlist, euclidean_distance};
 use bliss_audio::Song;
