+++ src/chroma.rs
@@ -365,7 +365,7 @@ fn chroma_stft(
 mod test {
     use super::*;
     #[cfg(feature = "ffmpeg")]
-    use crate::song::decoder::ffmpeg::FFmpeg as Decoder;
+    use crate::song::decoder::ffmpeg::FFmpegDecoder as Decoder;
     #[cfg(feature = "ffmpeg")]
     use crate::song::decoder::Decoder as DecoderTrait;
     #[cfg(feature = "ffmpeg")]
@@ -563,7 +563,7 @@ mod test {
 mod bench {
     extern crate test;
     use super::*;
-    use crate::song::decoder::ffmpeg::FFmpeg as Decoder;
+    use crate::song::decoder::ffmpeg::FFmpegDecoder as Decoder;
     use crate::song::decoder::Decoder as DecoderTrait;
     use crate::utils::stft;
     use crate::SAMPLE_RATE;
