+++ src/timbral.rs
@@ -259,7 +259,7 @@ impl Normalize for ZeroCrossingRateDesc {
 mod tests {
     use super::*;
     #[cfg(feature = "ffmpeg")]
-    use crate::song::decoder::ffmpeg::FFmpeg as Decoder;
+    use crate::song::decoder::ffmpeg::FFmpegDecoder as Decoder;
     #[cfg(feature = "ffmpeg")]
     use crate::song::decoder::Decoder as DecoderTrait;
     #[cfg(feature = "ffmpeg")]
